Change font metrics to all be integers, not floats.

+++ src/font/face.zig
@@ -39,23 +39,23 @@ pub const DesiredSize = struct {
 /// Metrics associated with the font that are useful for renderers to know.
 pub const Metrics = struct {
     /// Recommended cell width and height for a monospace grid using this font.
-    cell_width: f32,
-    cell_height: f32,
+    cell_width: u32,
+    cell_height: u32,
 
     /// For monospace grids, the recommended y-value from the bottom to set
     /// the baseline for font rendering. This is chosen so that things such
     /// as the bottom of a "g" or "y" do not drop below the cell.
-    cell_baseline: f32,
+    cell_baseline: u32,
 
     /// The position of the underline from the top of the cell and the
     /// thickness in pixels.
-    underline_position: f32,
-    underline_thickness: f32,
+    underline_position: u32,
+    underline_thickness: u32,
 
     /// The position and thickness of a strikethrough. Same units/style
     /// as the underline fields.
-    strikethrough_position: f32,
-    strikethrough_thickness: f32,
+    strikethrough_position: u32,
+    strikethrough_thickness: u32,
 };
 
 /// Additional options for rendering glyphs.
